fix: document select & sync remote w local
